+++ src/main/java/me/despical/commonsbox/item/ItemUtils.java
@@ -1,15 +1,30 @@
+/*
+ * CommonsBox - Library box of common utilities.
+ * Copyright (C) 2020 Despical
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package me.despical.commonsbox.item;
 
 import com.mojang.authlib.GameProfile;
 import com.mojang.authlib.properties.Property;
 
-import me.despical.commonsbox.compat.VersionResolver;
-import me.despical.commonsbox.compat.XMaterial;
-
 import java.lang.reflect.Field;
 import java.util.UUID;
 
-import org.bukkit.Material;
+import me.despical.commonsbox.compat.XMaterial;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.SkullMeta;
 
@@ -20,7 +35,11 @@ import org.bukkit.inventory.meta.SkullMeta;
  */
 public class ItemUtils {
 
-	public static final ItemStack PLAYER_HEAD_ITEM = VersionResolver.isCurrentLower(VersionResolver.ServerVersion.v1_13_R1) ? new ItemStack(Material.SKULL_ITEM, 1, (short) 3) : XMaterial.PLAYER_HEAD.parseItem();
+	public static final ItemStack PLAYER_HEAD_ITEM;
+
+	static {
+		PLAYER_HEAD_ITEM = XMaterial.PLAYER_HEAD.parseItem();
+	}
 
 	private ItemUtils() {}